+++ packages/dara-core/dara/core/auth/oidc/routes.py
@@ -21,7 +21,6 @@ import jwt
 from fastapi import Depends, HTTPException, Response
 
 from dara.core.auth.definitions import (
-    AUTH_COOKIE_KWARGS,
     BAD_REQUEST_ERROR,
     EXPIRED_TOKEN_ERROR,
     INVALID_TOKEN_ERROR,
@@ -29,7 +28,7 @@ from dara.core.auth.definitions import (
     SESSION_TOKEN_COOKIE_NAME,
 )
 from dara.core.auth.oidc.settings import OIDCSettings, get_oidc_settings
-from dara.core.auth.utils import get_cookie_expiration_from_token, sign_jwt
+from dara.core.auth.utils import set_cookie_from_token_expiration, sign_jwt
 from dara.core.http import post
 from dara.core.logging import dev_logger
 
@@ -131,20 +130,9 @@ async def sso_callback(
 
         # Set refresh token cookie if provided
         if oidc_tokens.refresh_token:
-            response.set_cookie(key=REFRESH_TOKEN_COOKIE_NAME, value=oidc_tokens.refresh_token, **AUTH_COOKIE_KWARGS)
+            set_cookie_from_token_expiration(response, REFRESH_TOKEN_COOKIE_NAME, oidc_tokens.refresh_token)
 
-        session_cookie_expiration = get_cookie_expiration_from_token(session_token)
-        if session_cookie_expiration is None:
-            response.set_cookie(key=SESSION_TOKEN_COOKIE_NAME, value=session_token, **AUTH_COOKIE_KWARGS)
-        else:
-            max_age, expires = session_cookie_expiration
-            response.set_cookie(
-                key=SESSION_TOKEN_COOKIE_NAME,
-                value=session_token,
-                max_age=max_age,
-                expires=expires,
-                **AUTH_COOKIE_KWARGS,
-            )
+        set_cookie_from_token_expiration(response, SESSION_TOKEN_COOKIE_NAME, session_token)
 
         return {'success': True}